requirements.txt
APScheduler==3.10.4
async-timeout==5.0.1
blinker==1.9.0
boto3==1.40.21
botocore==1.40.21
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
croniter==6.0.0
dnspython==2.7.0
email-validator==2.1.0
Flask==2.3.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.1.2
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
pillow==11.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
redis==6.4.0
requests==2.32.5
rq==2.5.0
s3transfer==0.13.1
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==1.26.20
Werkzeug==2.3.7
WTForms==3.1.0
zipp==3.23.0
